+++ src/Registration/pReg/Reg.py
@@ -21,8 +21,6 @@ Object-Oriented wrap for the cReg-to-Python interface pyreg.py
 
 import abc
 import sys
-import time
-import numbers
 import deprecation
 
 from pUtilities import *
@@ -31,12 +29,6 @@ import pyiutilities as pyiutil
 import pyreg
 
 import sirf.Reg_params as parms
-
-
-try:
-    input_ = raw_input
-except NameError:
-    pass
 
 if sys.version_info[0] >= 3 and sys.version_info[1] >= 4:
     ABC = abc.ABC
@@ -377,6 +369,30 @@ class NiftiImageData(SIRF.ImageData):
         for n in to_print:
             vec.push_back(n.handle)
         try_calling(pyreg.cReg_NiftiImageData_print_headers(vec.handle))
+
+    @staticmethod
+    def construct_from_complex_image(complex_im):
+        """Construct two NiftiImageData from a complex image"""
+        if not isinstance(complex_im, SIRF.ImageData):
+            raise AssertionError()
+        im_real = NiftiImageData()
+        im_imag = NiftiImageData()
+        im_real.handle = pyreg.cReg_NiftiImageData_from_complex_ImageData_real_component(complex_im.handle)
+        im_imag.handle = pyreg.cReg_NiftiImageData_from_complex_ImageData_imag_component(complex_im.handle)
+        return [im_real, im_imag]
+
+    @staticmethod
+    def are_equal_to_given_accuracy(im1, im2, accuracy):
+        """Check if two images match to a given accuracy"""
+        if not isinstance(im1, NiftiImageData) or not isinstance(im2, NiftiImageData):
+            raise AssertionError()
+        if im1.handle is None or im2.handle is None:
+            error('Cannot compare images as at least one is uninitialised')
+        h = pyreg.cReg_NiftiImageData_are_equal_to_given_accuracy(im1.handle, im2.handle, float(accuracy))
+        check_status(h, inspect.stack()[1])
+        value = pyiutil.intDataFromHandle(h)
+        pyiutil.deleteDataHandle(h)
+        return value
 
 
 class NiftiImageData3D(NiftiImageData):