+++ CHANGES.md
@@ -14,6 +14,7 @@
 * Inverse deformation images. Inverse displacements are also possible by converting to and from deformations.
 * NiftyPET projector wrapped (if STIR is built with NiftyPET)
 * Added `set_image_data_processor` to `PETAcquisitionModel`.  This allows for instance image-based PSF modelling.
+* Resampling of complex images.
 
 ## v2.1.0